move: 구현 완료한 기능을 test에서 main으로 이동
