+++ src/test/java/com/programmers/calculator/CalculatorTest.java
@@ -1,5 +1,6 @@
 package com.programmers.calculator;
 
+import com.programmers.domain.Calculator;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
