+++ src/test/java/com/programmers/postfixconverter/PostfixConverterTest.java
@@ -1,5 +1,6 @@
 package com.programmers.postfixconverter;
 
+import com.programmers.domain.PostfixConverter;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
